+++ drivers/mtd/stm32_hyperbus.c
@@ -189,7 +189,7 @@ static int stm32_hb_calibrate(struct stm32_hb_priv *priv)
 {
 	struct stm32_omi_plat *omi_plat = dev_get_plat(priv->omi_dev);
 	u32 prescaler;
-	u16 period_ps = 0;
+	u32 period_ps = 0;
 	u8 window_len = 0;
 	int ret;
 	bool bypass_mode = false;
