+++ include/stm32_omi.h
@@ -173,7 +173,7 @@ struct stm32_tap_window {
 };
 
 int stm32_omi_dlyb_configure(struct udevice *dev,
-			     bool bypass_mode, u16 period_ps);
+			     bool bypass_mode, u32 period_ps);
 int stm32_omi_dlyb_find_tap(struct udevice *dev, bool rx_only, u8 *window_len);
 int stm32_omi_dlyb_set_cr(struct udevice *dev, u32 dlyb_cr);
 void stm32_omi_dlyb_get_cr(struct udevice *dev, u32 *dlyb_cr);
